requirements.txt
# requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
jinja2
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
azure-storage-blob